Refactored base tables so they're internal. OrcaMDF now exposes metadata through the standard sys.* DMVs that you're used to from SQL Server. Subset exposed so far, more to come.

+++ src/OrcaMDF.Core/MetaData/SysrscolTIParser.cs
@@ -9,11 +9,11 @@ namespace OrcaMDF.Core.MetaData
 		public byte Precision;
 		public short MaxLength;
 		public short MaxInrowLength;
-		public int TypeID;
+		public byte TypeID;
 
 		public SysrscolTIParser(int ti)
 		{
-			TypeID = ti & 0xFF;
+			TypeID = (byte)(ti & 0xFF);
 
 			if (!Enum.IsDefined(typeof(SystemType), TypeID))
 				throw new ArgumentException("Unknown TypeID '" + TypeID + "'");
